requirements.txt
pywebview>=4.0
requests
openpyxl
pycryptodome
urllib3
pyinstaller
